Move DisjointSet to its own header file
